Add tests for number node
